Move include/excludeTags to Configuration level
These may have been included at the `SuiteConfiguration` level because
they are read next to other suite level configuration. At the place
where they are read (or anywhere in the `Runner`) the global
`Configuration` is available. Tags are currently only set at the global
level. If we add ability to configure tags using the URI syntax, we
likely will want to have a separate tracking for the global and suite
level.
