+++ pkgs/test/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.22.1-dev
+
 ## 1.22.0
 
 * Fix an issue with the github reporter where tests that fail asynchronously
