+++ pkgs/test/test/utils.dart
@@ -223,8 +223,6 @@ SuiteConfiguration suiteConfiguration(
         String? precompiledPath,
         Iterable<Pattern>? patterns,
         Iterable<RuntimeSelection>? runtimes,
-        BooleanSelector? includeTags,
-        BooleanSelector? excludeTags,
         Map<BooleanSelector, SuiteConfiguration>? tags,
         Map<PlatformSelector, SuiteConfiguration>? onPlatform,
         int? line,
@@ -249,8 +247,6 @@ SuiteConfiguration suiteConfiguration(
         precompiledPath: precompiledPath,
         patterns: patterns,
         runtimes: runtimes,
-        includeTags: includeTags,
-        excludeTags: excludeTags,
         tags: tags,
         onPlatform: onPlatform,
         line: line,
